Fix noon to make hour a known value rather than an implied value
Making "noon" consistent with "midnight".

This is important to me because (for some unknown reason) 12 is by default implied for hour.  So I have code to check to see if hour is "implied" vs "known", and I throw it out if it's implied and set it to zero.  But for this "noon" I was throwing it out, which is not desired.

An alternative solution for me would be to allow a default hour to be set in config rather than implying 12.

Should the minute/second/milliseconds also be known for both midnight and noon?  Doesn't a